move succeeded
