--- site/activity-react/components/modal/modal.js
@@ -1,12 +1,0 @@
-import * as React from 'react';
-import './style.scss';
-
-export default class Modal extends React.Component {
-    render() {
-        return (
-            <div> 
-                {this.props.children}
-            </div>
-        )
-    }
-}
