--- site/activity-react/components/modal/style.scss
@@ -1,24 +1,0 @@
-#modal-root {
-    position: relative;
-    z-index: 999;
-  }
-  
-  .app {
-    height: 10em;
-    width: 10em;
-    background: lightblue;
-    overflow: hidden;
-  }
-  
-  .modal {
-    background-color: rgba(0,0,0,0.5);
-    position: fixed;
-    height: 100%;
-    width: 100%;
-    top: 0;
-    left: 0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
