+++ site/activity-react/pages/container/files/index.js
@@ -1,10 +1,167 @@
 import * as React from 'react';
 import './style.scss'
 
+var ReactDOM = require('react-dom')
 import { timeBefore, sortByCreateTime, formatDate } from '../../../utils/util'
 import api from '../../../utils/api';
 import Page from '../../../components/page'
 import fileUploader from '../../../utils/file-uploader';
+
+
+
+const root = document.getElementById('app')
+
+class Modal extends React.Component {
+    componentDidMount = async () => {
+        this.teamId = this.props.teamId
+        this.folderId = this.props.folderId
+        this.initDirList()
+        this.initTeamInfo()
+        this.initTeamFile()
+    }
+    state = {
+        teamInfo : {},
+        fileList: [],
+        dirList: [],
+    }
+
+    initDirList = () => {
+        if(!this.curDir) {
+            this.curDir = this.props.dir || '/'
+        }
+        if(this.curDir == '/') { 
+            this.setState({
+                dirList: []
+            })
+            return
+        }
+        let splitDir = this.curDir.split('/')
+        let totalDirList = []
+        splitDir.map((item, idx) => {
+            if(idx == 0) {
+                totalDirList.push({
+                    name: '根目录',
+                    dir: ''
+                })
+            } else {
+                totalDirList.push({
+                    name: item,
+                    dir: totalDirList[idx - 1].dir + '/' + item
+                }) 
+            }
+        })
+        totalDirList[0].dir = '/'
+        this.setState({
+            dirList: totalDirList
+        })
+    }
+
+    initTeamInfo = async () => {
+        const result = await api('/api/team/info', {
+            method: 'POST',
+            body: {
+                teamId: this.teamId
+            }
+        })
+        if(result.data) {
+            this.setState({
+                teamInfo: result.data
+            })
+        }
+    }
+
+    initTeamFile = async () => {
+        const result = await api('/api/file/getDirFileList', {
+            method: 'POST',
+            body: {
+                dirInfo: {
+                    teamId: this.teamId,
+                    dir: this.curDir,
+                }
+            }
+        })
+        if (result && result.data && result.data.fileList) {
+            this.setState({
+                fileList: result.data.fileList
+            })
+        }
+    }
+
+    folderClickHandle = (folderName) => {
+        const tarDir = this.curDir + (this.curDir == '/' ? '' : '/') + folderName
+        this.curDir = tarDir
+        this.initDirList()
+        this.initTeamFile()
+
+    }
+
+    headDirClickHandle = (dir) => {
+        if(dir == '/') {
+            this.curDir = '/'
+            this.initDirList()
+            this.initTeamFile()
+        } else {
+            this.curDir = dir
+            this.initDirList()
+            this.initTeamFile()
+        }
+    }
+
+    closeWindow = () => {
+        this.props.callbackParent('')
+    }
+
+    confirm = () => {
+        this.props.callbackParent(this.curDir)
+    }
+
+    render() {
+        console.log(this.state.fileList)
+        console.log(this.state.dirList)
+        return (
+            <div className="window" >
+                <div className="outerBox">
+                    <Page className="move-File">
+                        <div className="file-con">
+                        <div> 移动到： </div>
+                            <div className="file-dir">
+                                {
+                                    this.state.dirList.length ?
+                                        <div>
+                                            {
+                                                this.state.dirList.map((item, idx) => (
+                                                    <span key={"dir-list-" + idx} onClick={() => { this.headDirClickHandle(item.dir) }}>{item.name} {idx == this.state.dirList.length - 1 ? '' : '>'} </span>
+                                                ))
+                                            }
+                                        </div>
+                                        : '根目录'
+                                }
+                            </div>
+                            <div className="file-list">
+                                {
+                                    this.state.fileList.map((item, idx) => {
+
+                                        if (item.fileType == 'folder' && item._id != this.folderId) {
+                                            return (
+                                                <div className="file-line files" key={item.fileType + '-' + item._id}>
+                                                    <div className="name" onClick={() => { this.folderClickHandle(item.name) }}>{item.name}</div>
+                                                </div>
+                                            )
+                                        }
+
+
+                                    })
+                                }
+                            </div>
+                            <div className="btn" onClick={this.confirm}> confirm </div>
+                            <div className="btn" onClick={this.closeWindow}> close </div>
+                        </div>
+                    </Page>
+                </div>
+            </div>
+        )
+    }
+}
 
 export default class Files extends React.Component {
     componentDidMount = async () => {
@@ -24,6 +181,8 @@ export default class Files extends React.Component {
         renameName: '',
 
         dirList: [],
+        modal: document.createElement('div'),
+        moveItem: '',
     }
 
     initDirList = () => {
@@ -136,6 +295,63 @@ export default class Files extends React.Component {
             this.state.dir = path;
         }
         this.getDirFileListHandle()
+    }
+
+    moveHandle = async (item,tarDir) => {
+        console.log(this.state.moveItem)
+        if(item.fileType == 'file') {
+            const result = await api('/api/file/moveFile', {
+                method: 'POST',
+                body: {
+                    fileInfo: {
+                        teamId: this.teamId,
+                        dir: this.curDir,
+                        fileName: this.state.moveItem.name,
+                        tarDir: tarDir,
+                    }
+                }   
+            })
+            
+            if(result.state.code == 0) {
+                window.toast("移动文件成功")
+            } else {
+                window.toast(result.state.msg)
+            }
+        }
+        else {
+            const result = await api('/api/file/moveFolder', {
+                method: 'POST',
+                body: {
+                    folderInfo: {
+                        teamId: this.teamId,
+                        dir: this.curDir,
+                        folderName: this.state.moveItem.name,
+                        tarDir: tarDir,
+                    }
+                }
+            })
+ 
+            if(result.state.code == 0) {
+                window.toast("移动文件夹成功")
+            } else {
+                window.toast(result.state.msg)
+            }
+        }
+        this.initTeamFile()
+    }
+
+    onChildChanged = (moveTarDir) => {
+        if(moveTarDir != '') {
+            this.moveHandle(this.state.moveItem,moveTarDir)
+        }
+        document.getElementById('app').removeChild(this.state.modal)
+        this.state.moveItem = ''
+    }
+
+    openMoveModalHandle = (item) => {
+        this.state.moveItem = item
+        ReactDOM.render(<Modal teamId={this.teamId} folderId={item._id} callbackParent={this.onChildChanged} />, this.state.modal)
+        document.getElementById('app').appendChild(this.state.modal)
     }
 
     uploadFileHandle = async (e) => {
@@ -374,7 +590,7 @@ export default class Files extends React.Component {
                                                 <div className="size">-</div>
                                                 <div className="last-modify">{formatDate(item.last_modify_time)}</div>
                                                 <div className="tools">
-                                                    <span>移动</span>
+                                                    <span onClick={() => { this.openMoveModalHandle(item)}}>移动</span>
                                                     <span onClick={() => { this.renameHandle(item)}}> 重命名 </span> 
                                                     <span onClick={() => { this.deleteHandle('folder', item.name) }}>删除</span>
                                                 </div>
@@ -389,7 +605,7 @@ export default class Files extends React.Component {
                                                 <div className="last-modify">{formatDate(item.last_modify_time)}</div>
                                                 <div className="tools">
                                                     <span onClick={() => { this.downloadHandle(item.ossKey) }}>下载</span>
-                                                    <span>移动</span>
+                                                    <span onClick={() => { this.openMoveModalHandle(item)}}>移动</span>
                                                     <span onClick={() => { this.renameHandle(item)}}> 重命名 </span> 
                                                     <span onClick={() => { this.deleteHandle('file', item.name) }}>删除</span>
                                                 </div>
